HyperLogLog-based anonymous hashing of IP addresses
Julia implementation of HyperLogLog hashing of IP addresses
(both IPv4 and IPv6) using format-preserving encryption with a Feistel
cipher using SHA512 as the round function. This implementation generates
a hex string that is in the format that Snowflake uses for HyperLogLog
aggregation, which means that this data can be fed into Snowflake,
convert it to binary from hex, and then you can use the built-in
`hll_combine` aggregation function to build up a HyperLogLog sketch of
any collection of hashes.  The sketch can be converted to a cardinality
estimate with expected average error around 1.6% by calling
`hll_estimate`. This seems to work well in practice on actual
collections of IP addresses.

The hash function is designed so that we can estimate collections of up
to 1 billion IP addresses, yet it is guaranteed that no IP addresses can
be uniquely identified by their hash value since at least 4 IP addresses
map to the rarest hash values and most IP addresses map to hash values
that are shared with billions of other IP addresses. Hash values are
only 16.25 bits and only 12 bits of information are stored for typical IP
addresses. This hashing technique protects anonymity while also
allowing us to accurately estimate the number of unique IP addresses in
arbitrary traffic traces.

This implementation was adapted from the original, which contains a java
implementation as well:
https://github.com/StefanKarpinski/HyperLogLogHashIPs.jl